Refactor aruba spec setup.
This will facilitate writing other specs in using aruba.
