+++ spec/integration/order_spec.rb
@@ -1,6 +1,7 @@
-RSpec.describe 'command line', :ui, :slow do
-  let(:stderr) { StringIO.new }
-  let(:stdout) { StringIO.new }
+require 'support/aruba_support'
+
+RSpec.describe 'command line', :ui do
+  include_context "aruba support"
 
   before :all do
     write_file 'spec/simple_spec.rb', """
@@ -199,13 +200,5 @@ RSpec.describe 'command line', :ui, :slow do
   def split_in_half(array)
     length, midpoint = array.length, array.length / 2
     return array.slice(0, midpoint), array.slice(midpoint, length)
-  end
-
-  def run_command(cmd)
-    in_current_dir do
-      RSpec::Core::Runner.run(cmd.split, stderr, stdout)
-    end
-  ensure
-    RSpec.reset
   end
 end
